refactor：storybook：component：TextField：関数切り出し
